refactor(debates): give "is this a debate" one home, shared with the entity route
The predicate landed in an explore module, which was the wrong place for it.
The entity route was already asking the same question inline to decide whether
to render the full-screen debates feed instead of a value sheet — so this
change had put a second spelling of one decision at the other end of the
journey it enables. Explore says "this is a debate, navigate there" and the
route decides whether to give them a debate; if the two ever disagreed, the
symptom is a reader clicking a debate and landing on a page that does not think
it is one.

Not a live bug: the API returns type ids hyphenless, so the route's raw
equality matches today and the normalization here is belt and braces. It is a
bug waiting for the first id that arrives in the other spelling.

So the predicate moves to core/debates, alongside the other small
single-purpose modules there, and all three callers read it. It accepts
undefined because the route asks off a fetch that can come back empty, the same
shape the Entities helpers use for relations.
