+++ apps/web/app/space/(entity)/[id]/[entityId]/page.tsx
@@ -5,7 +5,7 @@ import { notFound } from 'next/navigation';
 import { EVENT_SCHEMA } from '~/core/community-calls/constants';
 import { getRecordingUrls } from '~/core/community-calls/recordings';
 import { DebateEntityView } from '~/core/debates/browse/debate-entity-view';
-import { DEBATE_TYPE_ID } from '~/core/debates/ontology';
+import { isDebateEntity } from '~/core/debates/is-debate-entity';
 import { entityHasOnlyPostType } from '~/core/utils/entity/entities';
 
 import { CommunityCallRecording } from '~/partials/community-calls/community-call-recording';
@@ -40,7 +40,7 @@ export default async function EntityTemplateStrategy(props: Props) {
 
   // A Debate is a live video, not a value sheet: browse mode drops you into the debates feed
   // anchored to this debate, and the raw entity page is reserved for edit mode.
-  if (result?.entity?.types.some(t => t.id === DEBATE_TYPE_ID)) {
+  if (isDebateEntity(result?.entity?.types)) {
     return (
       <DebateEntityView
         spaceId={params.id}
